requirements.txt
streamlit==1.44.1
openai==1.70.0
pillow==11.1.0
